requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
frozenlist==1.5.0
h11==0.14.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
magic-filter==1.0.12
multidict==6.1.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.17.1
